third_party/WebKit/LayoutTests/platform/mac-mac10.11/external/wpt/css/css-animations/Element-getAnimations.tentative-expected.txt

This is a testharness.js-based test.
PASS getAnimations for non-animated content
PASS getAnimations for CSS Animations
FAIL getAnimations returns CSSAnimation objects for CSS Animations assert_equals: Interface of returned animation is CSSAnimation expected "[object CSSAnimation]" but got "[object Animation]"
PASS getAnimations for multi-property animations
FAIL getAnimations for both CSS Animations and CSS Transitions at once assert_equals: First-returned animation is the CSS Transition expected "[object CSSTransition]" but got "[object Animation]"
PASS getAnimations for CSS Animations that have finished
PASS getAnimations for CSS Animations that have finished but are forwards filling
PASS getAnimations for CSS Animations with animation-name: none
PASS getAnimations for CSS Animations with animation-name: missing
FAIL getAnimations for CSS Animations where the @keyframes rule is added later assert_true: Newly added animation has a later start time expected true got false
PASS getAnimations for CSS Animations with duplicated animation-name
PASS getAnimations for CSS Animations with empty keyframes rule
PASS getAnimations for CSS animations in delay phase
PASS getAnimations for zero-duration CSS Animations
FAIL getAnimations returns objects with the same identity assert_equals: animation's play state is updated expected "paused" but got "pending"
PASS getAnimations for CSS Animations that are canceled
FAIL getAnimations for CSS Animations follows animation-name order assert_equals: animation order after prepending to list expected (string) "anim1" but got (undefined) undefined
PASS { subtree: false } on a leaf element returns the element's animations and ignore pseudo-elements
FAIL { subtree: true } on a leaf element returns the element's animations and its pseudo-elements' animations assert_equals: getAnimations({ subtree: true }) should return animations on pseudo-elements expected 3 but got 1
PASS { subtree: false } on an element with a child returns only the element's animations
FAIL { subtree: true } on an element with a child returns animations from the element, its pseudo-elements, its child and its child pseudo-elements assert_equals: Should find all elements, pesudo-elements that parent has expected 6 but got 1
FAIL { subtree: true } on an element with many descendants returns animations from all the descendants assert_equals: Should find all descendants of the element expected 5 but got 1
Harness: the test ran to completion.
